refactor : 길안내 시작시 기존 레디스 세션 삭제 후 생성, dto 및 파라미터 타입 변경

See merge request s13-final/S13P31A305!155

+++ backend/ongil/src/main/java/kr/co/ongil/domain/navigation/dto/request/EndNavigationRequest.java
@@ -7,11 +7,11 @@ public record EndNavigationRequest (
 
     @NotNull
     @Schema(description = "환자 ID", example = "1")
-    Long patientId,
+    Integer patientId,
 
     @NotNull
     @Schema(description = "네비게이션 ID", example = "1")
-    Long navigationId,
+    Integer navigationId,
 
     @Schema(description = "정상 종료 여부", example = "true")
     Boolean isSuccessful
